pyproject.toml
[tool.poetry]
name = "Render Docker Test"
version = "0.1.0"
description = ""
authors = ["Suiketsu Iranai <nvrmndr.desu@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = {version = "^20.2", extras = ["job-queue", "webhooks"]}
uvloop = {version = "^0.17.0", platform = "linux"}
snscrape = "^0.6.2.20230320"
orjson = "^3.8.9"
pyexiftool = "^0.5.5"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
ruff = "^0.0.260"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
profile = "black"
skip = ["alembic.ini"]
skip_glob = [".*/", ".*", "log/", "*.log"]
line_length = 90
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
lines_between_types = 1

[tool.black]
line-length = 90
include = '\.pyi?$'

[tool.ruff]
select = ["E", "F", "B"]
ignore = []
per-file-ignores = {}
line-length = 90
target-version = "py311"

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
